refactor: move Git logic to separate class
